Use a REMOVED event when a file is removed from a catalog
If a file is deleted from disk it must be deleted from the catalogs
as well, but if a file is removed from a catalog it's not deleted
from disk.  For this reason we need to differentiate
the two cases using REMOVED when removing a file from a catalog,
and DELETE when deleting from disk.