[importer] use a single import time when creating subfolders
while gthumb-2.10 calculated a folder name once for an import, the new
gthumb calculates the destination for every file individually. This
can lead to possibly unexpected behavior if you use "current time" and
"custom" with an format like the old one (%Y-%m-%d--%H.%M.%S), because
you now get a new folder every second... The attached patch tries to
change this (hopefully in a sane way...) by setting an
import_start_time at the beginning of the import and using this as the
"current time" for the complete import.