Give priority to gvfs preview icons over exiv2
As described by Michael Chudobiak in the bug report:
Check for preview::icon first (from the gphoto backend, for
example), then use exiv2 to read the embedded thumbnail, and lastly
try the whole file.
This is important for the importer - the gphoto gvfs backend provides
the fastest access to the thumbnails (much faster than exiv2).

[bug #616518]