activate commands on key release with the exception of F5
F5 cannot be activated on key release because it's used to
activate the slideshow as well, and the release key event
relative to the slideshow activation is captured by the 
fullscreen window that quits soon after.

The other commands are activated on key release to avoid
double activation of a command on long key presses.