correctly display images with wrong extension name
If after reading the metadata we got a different mime type and the current
viewer is the default file viewer it's likely that we have an image with a
wrong extension.  Try to load the file again to see if the mime type can be
viewed by a different viewer_page.

[bug fixed]