build: Fix parallel build problems with enum generation
If the enum rule is triggered because one of the dependencies changed, but the
resulting enum-types file is identical, it isn't copied across.  This causes
problems in parallel builds because make busy-loops on the file changing instead
of just running the commands sequentially.

The upshot is that a make -j where this happens (simply touching a .h file will
do it) will sit spinning inside make.

Fix this by pruning the dependencies to only those actually required, and doing
a mv instead of cmp||cp.  If the headers have been modified then we'll be
rebuilding anyway so the optimisation of the cmp is simply causing bugs.

https://bugzilla.gnome.org/show_bug.cgi?id=680925