meson: Add a workaround for extension files translation
`extension` files uses the `Desktop Entry Spec`[0] to describe the
extensions. They also use a subset of the supported keywords by
gettext. These allow the `extension` files to be properly
translated.

However, they are not using the `.desktop` extension, which is used
by gettext to detect the type of the files, so it fallbacks to `C`
and does not extract the strings to be translated.

[0] https://www.freedesktop.org/wiki/Specifications/desktop-entry-spec/