Added mnemonics to file manager and browser context menu
Currently the context menu of the file manager and the browser do not
have any mnemonics, making accessibility via keyboard difficult.
With these adjustment, the context menu is adjusted to show mnemonics
for all menu entries.
Additionally, a mnemonic is introduced into the Move to / Copy to dialog
to be able to initiate the Move or Copy via keyboard.

Where possible the following standards for mnemonics have been followed:
https://developer.gnome.org/hig-book/unstable/input-keyboard.html.en