Video thumbnailer: use the default thumbnail size if invalid
From the merge request
https://gitlab.gnome.org/GNOME/gthumb/-/merge_requests/46

The current webalbum exporter code specifies a thumbnail size of "-1"
and the video thumbnailer uses this to create a single-pixel-sized
thumbnail. This patch checks for this crazy size, and sets it to a
reasonable default, instead.

Patch by Linas Vepstas

Patch changed because the merge request is old and it is not valid
anymore.